Fixed regress bugs #3

+++ app/src/main/java/info/tduty/typetalk/utils/Utils.kt
@@ -4,6 +4,6 @@ object Utils {
 
     fun getSuccessCompletedTaskPercent(countTask: Int, countSuccessTask: Int) : Double {
         if (countTask <= 0) return 0.0
-        return kotlin.math.ceil((countSuccessTask / countTask) * 100.0)
+        return kotlin.math.ceil((countSuccessTask.toFloat() / countTask.toFloat()) * 100.0)
     }
 }
